requirements.txt
absl-py==2.2.2
aioice==0.10.1
aiortc==1.11.0
altair==5.5.0
altgraph==0.17.4
astunparse==1.6.3
attrs==25.3.0
av==14.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.10
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dnspython==2.7.0
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.55.3
fsspec==2025.2.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.1
google-pasta==0.2.0
grpcio==1.71.0
gTTS==2.5.4
h5py==3.13.0
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.2
kiwisolver==1.4.8
labelImg==1.8.6
lap==0.5.12
libclang==18.1.1
lxml==5.3.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.9
narwhals==1.33.0
networkx==3.4.2
numpy==1.26.2
onnx==1.17.0
onnxruntime==1.20.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.15.0
packaging==24.1
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
protobuf==4.25.7
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
pyee==13.0.0
pygame==2.6.1
Pygments==2.19.1
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pylibsrtp==0.12.0
pynput==1.7.7
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2025.1
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==75.2.0
six==1.16.0
smmap==5.0.2
sounddevice==0.5.1
streamlit==1.44.1
streamlit-autorefresh==1.0.1
streamlit-webrtc==0.62.4
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.77
ultralytics-thop==2.0.14
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
